+++ .storybook/main.ts
@@ -7,6 +7,15 @@ const config: StorybookConfig = {
     name: '@storybook/react-vite',
     options: {},
   },
+  async viteFinal(config) {
+    return {
+      ...config,
+      optimizeDeps: {
+        ...config.optimizeDeps,
+        include: [...(config.optimizeDeps?.include ?? []), 'antd'],
+      },
+    }
+  },
 }
 
 export default config