+++ src/components/DatePicker/index.ts
@@ -1,2 +1,2 @@
 export { DatePicker } from './DatePicker'
-export type { DatePickerProps, DatePickerShowTime } from './DatePicker'
+export type { DatePickerProps, DatePickerShowTime, DatePickerInputProps } from './DatePicker'